+++ generators/r1279_shared.c
@@ -52,7 +52,7 @@ static void *create(const CallerAPI *intf)
     // pcg_rxs_m_xs64 for initialization
     uint64_t state = intf->get_seed64();
     for (size_t k = 1; k <= RGEN_A; k++) {    
-        obj->x[k] = pcg_bits64(&state);
+        obj->x[k] = (uint32_t) pcg_bits64(&state);
     }
     obj->i = RGEN_A; obj->j = RGEN_B;
     return (void *) obj;
